+++ packages/core/src/index.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-/* Graph mixins */
+/* Graph mixins - side effects */
 import './view/mixins/PortsMixin';
 import './view/mixins/PanningMixin';
 import './view/mixins/ZoomMixin';
@@ -39,11 +39,10 @@ import './view/mixins/PageBreaksMixin';
 import './view/mixins/GroupingMixin';
 import './view/mixins/OrderMixin';
 
-// GraphCodec
 export { Graph } from './view/Graph';
 
-export { GraphDataModel, ModelCodec } from './view/GraphDataModel';
-export { GraphView, GraphViewCodec } from './view/GraphView';
+export { GraphDataModel } from './view/GraphDataModel';
+export { GraphView } from './view/GraphView';
 export { default as LayoutManager } from './view/layout/LayoutManager';
 export { default as Outline } from './view/other/Outline';
 export { default as PrintPreview } from './view/other/PrintPreview';
@@ -51,28 +50,21 @@ export { default as SwimlaneManager } from './view/layout/SwimlaneManager';
 export { default as Client } from './Client';
 
 export { default as CellAttributeChange } from './view/undoable_changes/CellAttributeChange';
-export { ChildChange, ChildChangeCodec } from './view/undoable_changes/ChildChange';
+export { ChildChange } from './view/undoable_changes/ChildChange';
 export { default as CollapseChange } from './view/undoable_changes/CollapseChange';
 export { default as CurrentRootChange } from './view/undoable_changes/CurrentRootChange';
 export { default as GeometryChange } from './view/undoable_changes/GeometryChange';
-export { RootChange, RootChangeCodec } from './view/undoable_changes/RootChange';
+export { RootChange } from './view/undoable_changes/RootChange';
 export { default as SelectionChange } from './view/undoable_changes/SelectionChange';
 export { default as StyleChange } from './view/undoable_changes/StyleChange';
-export {
-  TerminalChange,
-  TerminalChangeCodec,
-} from './view/undoable_changes/TerminalChange';
+export { TerminalChange } from './view/undoable_changes/TerminalChange';
 export { default as ValueChange } from './view/undoable_changes/ValueChange';
 export { default as VisibleChange } from './view/undoable_changes/VisibleChange';
 
-export { EditorKeyHandler, EditorKeyHandlerCodec } from './editor/EditorKeyHandler';
-export { EditorPopupMenu, EditorPopupMenuCodec } from './editor/EditorPopupMenu';
-// TODO restore EditorToolbarCodec if needed
-// currently, the following error occurs when running vitejs build
-// [ERROR] No matching export in "../core/dist/esm/editor/EditorToolbar.js" for import "EditorToolbarCodec"
-// export { EditorToolbar, EditorToolbarCodec } from './editor/EditorToolbar';
+export { EditorKeyHandler } from './editor/EditorKeyHandler';
+export { EditorPopupMenu } from './editor/EditorPopupMenu';
 export { EditorToolbar } from './editor/EditorToolbar';
-export { Editor, EditorCodec } from './editor/Editor';
+export { Editor } from './editor/Editor';
 
 export { default as CellHighlight } from './view/cell/CellHighlight';
 export { default as CellMarker } from './view/cell/CellMarker';
@@ -121,8 +113,10 @@ export { default as SwimlaneOrdering } from './view/layout/hierarchical/Swimlane
 
 export { default as Codec } from './serialization/Codec';
 export { default as CodecRegistry } from './serialization/CodecRegistry';
-export { default as GenericChangeCodec } from './view/undoable_changes/GenericChangeCodec';
 export { default as ObjectCodec } from './serialization/ObjectCodec';
+export * from './serialization/ModelXmlSerializer';
+export * from './serialization/codecs';
+export * from './serialization/register';
 
 export { default as ActorShape } from './view/geometry/ActorShape';
 export { default as LabelShape } from './view/geometry/node/LabelShape';
@@ -180,10 +174,6 @@ export { default as Rectangle } from './view/geometry/Rectangle';
 export { default as EdgeStyle } from './view/style/EdgeStyle';
 export { default as Perimeter } from './view/style/Perimeter';
 export { default as StyleRegistry } from './view/style/StyleRegistry';
-// TODO restore StylesheetCodec if needed
-// currently, the following error occurs when running vitejs build
-// [ERROR] No matching export in "../core/dist/esm/view/style/Stylesheet.js" for import "StylesheetCodec"
-// export { Stylesheet, StylesheetCodec } from './view/style/Stylesheet';
 export { Stylesheet } from './view/style/Stylesheet';
 
 export * as DomHelpers from './util/domHelpers';
@@ -217,7 +207,7 @@ export { default as Clipboard } from './util/Clipboard';
 export { default as UndoableEdit } from './view/undoable_changes/UndoableEdit';
 export { default as UndoManager } from './view/undoable_changes/UndoManager';
 
-export { Cell, CellCodec } from './view/cell/Cell';
+export { Cell } from './view/cell/Cell';
 export { default as CellEditorHandler } from './view/handler/CellEditorHandler';
 export { default as CellOverlay } from './view/cell/CellOverlay';
 export { default as CellPath } from './view/cell/CellPath';