+++ packages/core/src/view/handler/CellEditorHandler.ts
@@ -712,10 +712,7 @@ class CellEditorHandler implements GraphPlugin {
     }
 
     const tooltipHandler = this.graph.getPlugin('TooltipHandler') as TooltipHandler;
-
-    if (tooltipHandler) {
-      tooltipHandler.hideTooltip();
-    }
+    tooltipHandler?.hideTooltip();
 
     const state = this.graph.getView().getState(cell);
 
@@ -801,8 +798,7 @@ class CellEditorHandler implements GraphPlugin {
       // Workaround for initial offsetHeight not ready for heading in markup
       if (
         this.autoSize &&
-        // @ts-ignore
-        (this.graph.model.isEdge(state.cell) || state.style.overflow !== 'fill')
+        (state.cell.isEdge() || state.style.overflow !== 'fill')
       ) {
         window.setTimeout(() => {
           this.resize();