+++ packages/html/stories/Labels.stories.js
@@ -22,17 +22,24 @@ import {
   constants,
   Rectangle,
 } from '@maxgraph/core';
-
-import { globalTypes } from '../.storybook/preview';
+import {
+  globalTypes,
+  globalValues,
+  rubberBandTypes,
+  rubberBandValues,
+} from './shared/args.js';
+// style required by RubberBand
+import '@maxgraph/core/css/common.css';
 
 export default {
   title: 'Labels/Labels',
   argTypes: {
     ...globalTypes,
-    rubberBand: {
-      type: 'boolean',
-      defaultValue: true,
-    },
+    ...rubberBandTypes,
+  },
+  args: {
+    ...globalValues,
+    ...rubberBandValues,
   },
 };
 
@@ -60,7 +67,7 @@ const Template = ({ label, ...args }) => {
   // Do not allow removing labels from parents
   graphHandler.removeCellsFromParent = false;
 
-  // Autosize labels on insert where autosize=1
+  // Allow auto-size labels on insert
   graph.autoSizeCellsOnAdd = true;
 
   // Allows moving of relative cells