+++ packages/html/stories/Overlays.stories.js
@@ -23,13 +23,15 @@ import {
   utils,
   ImageBox,
 } from '@maxgraph/core';
-
-import { globalTypes } from '../.storybook/preview';
+import { globalTypes, globalValues } from './shared/args.js';
 
 export default {
   title: 'Effects/Overlays',
   argTypes: {
     ...globalTypes,
+  },
+  args: {
+    ...globalValues,
   },
 };
 
@@ -54,6 +56,14 @@ const Template = ({ label, ...args }) => {
   // Enables tooltips for the overlays
   graph.setTooltips(true);
 
+  function pickAlignValueRandomly() {
+    return ['left', 'center', 'right'][Math.floor(Math.random() * 3)];
+  }
+
+  function pickVerticalAlignValueRandomly() {
+    return ['top', 'bottom'][Math.floor(Math.random() * 2)];
+  }
+
   // Installs a handler for click events in the graph
   // that toggles the overlay for the respective cell
   graph.addListener(InternalEvent.CLICK, (sender, evt) => {
@@ -61,17 +71,18 @@ const Template = ({ label, ...args }) => {
 
     if (cell != null) {
       const overlays = graph.getCellOverlays(cell);
-
-      if (overlays.length == 0) {
+      if (overlays.length === 0) {
         // Creates a new overlay with an image and a tooltip
         const overlay = new CellOverlay(
           new ImageBox('/images/check.png', 16, 16),
-          'Overlay tooltip'
+          'Overlay tooltip',
+          pickAlignValueRandomly(),
+          pickVerticalAlignValueRandomly()
         );
 
         // Installs a handler for clicks on the overlay
         overlay.addListener(InternalEvent.CLICK, (sender, evt2) => {
-          utils.alert('Overlay clicked');
+          window.alert('Overlay clicked');
         });
 
         // Sets the overlay for the cell in the graph
@@ -104,7 +115,7 @@ const Template = ({ label, ...args }) => {
     });
     const v2 = graph.insertVertex({
       parent,
-      value: 'Doubleclick',
+      value: 'Double Click',
       position: [200, 150],
       size: [100, 40],
     });